cloud_function/requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-bigquery
google-genai
PyMuPDF
Pillow
numpy
rdkit
pillow
python-docx